bug fixed.

+++ ksf.c
@@ -1,5 +1,5 @@
 /*     ksf.c (Kanji Statistics Filter)     
- *     Time-stamp: "2019-01-11 13:04:52 yamagen"
+ *     Time-stamp: "2019-01-11 13:21:06 yamagen"
  *     Hilofumi Yamamoto <yamagen@ila.titech.ac.jp>
  *
  *     Hilofumi Yamamoto 03-July-1998 <yamagen@ucsd.edu>
@@ -369,18 +369,18 @@ dump_kanji_stat (FILE *output, unsigned long *kanji_array,stats_record *stats)
     cumulative_percentage = running_total * percentage_factor;
     
     if(OBOSOLETE)
-      fprintf(output,"%d\t%X%X\t%c%c\t%d\t%9.5f\n",
-	      (int)kanji_number++,
+      fprintf(output,"%ld\t%X%X\t%c%c\t%ld\t%9.5f\n",
+	      kanji_number++,
 	      vector_entry->U,	vector_entry->L,
 	      vector_entry->U,	vector_entry->L,
-	      (int)vector_entry->count,
+	      vector_entry->count,
 	      cumulative_percentage);
 
     if(CNT)
-      fprintf(output,"%d\t%c%c\t%d\t%6.4f\n",
-	      (int)kanji_number++,
+      fprintf(output,"%ld\t%c%c\t%ld\t%6.4f\n",
+	      kanji_number++,
 	      vector_entry->U,	vector_entry->L,
-	      (int)vector_entry->count,
+	      vector_entry->count,
 	      cumulative_percentage);
 
     if (vector_entry->count == 1)
@@ -392,21 +392,21 @@ dump_kanji_stat (FILE *output, unsigned long *kanji_array,stats_record *stats)
     fprintf(output,"\n");
   total_chars = 100.0 / stats->jis_cnt;
   if(TAB){
-    fprintf(output,"files        %9d\n", (int)stats->success_files);
-    fprintf(output,"kanji(total) %9d\t(%4.1f%%)\n", (int)stats->kanji_cnt, stats->kanji_cnt*total_chars);
-    fprintf(output,"kanji(unique)%9d\n", (int)stats->unique_kanji);
-    fprintf(output,"kanji(once)  %9d\t(%4.1f%% of unique)\n", (int)single_total, single_total*100.0/stats->unique_kanji);
-    fprintf(output,"hiragana     %9d\t(%4.1f%%)\n", (int)stats->hiragana_cnt, stats->hiragana_cnt*total_chars);
-    fprintf(output,"katakana     %9d\t(%4.1f%%)\n", (int)stats->katakana_cnt, stats->katakana_cnt*total_chars);
-    fprintf(output,"Arabic       %9d\t(%4.1f%%)\n", (int)stats->number_cnt, stats->number_cnt*total_chars);
-    fprintf(output,"Roman        %9d\t(%4.1f%%)\n", (int)stats->roman_cnt, stats->roman_cnt*total_chars);
-    fprintf(output,"Cyrillic     %9d\t(%4.1f%%)\n", (int)stats->cyrillic_cnt, stats->cyrillic_cnt*total_chars);
-    fprintf(output,"Greek        %9d\t(%4.1f%%)\n", (int)stats->greek_cnt, stats->greek_cnt*total_chars);
-    fprintf(output,"graphic      %9d\t(%4.1f%%)\n", (int)stats->graphic_cnt, stats->graphic_cnt*total_chars);
-    fprintf(output,"punctuation  %9d\t(%4.1f%%)\n", (int)stats->punctuation_cnt, stats->punctuation_cnt*total_chars);
-    fprintf(output,"symbols      %9d\t(%4.1f%%)\n", (int)stats->symbol_cnt, stats->symbol_cnt*total_chars);
-    fprintf(output,"space        %9d\t(%4.1f%%)\n", (int)stats->space_cnt, stats->space_cnt*total_chars);
-    fprintf(output,"other        %9d\t(%4.1f%%)\n", (int)stats->other_cnt, stats->other_cnt*total_chars);
+    fprintf(output,"files        %9ld\n", stats->success_files);
+    fprintf(output,"kanji(total) %9ld\t(%4.1f%%)\n", stats->kanji_cnt, stats->kanji_cnt*total_chars);
+    fprintf(output,"kanji(unique)%9ld\n",            stats->unique_kanji);
+    fprintf(output,"kanji(once)  %9ld\t(%4.1f%% of unique)\n", single_total, single_total*100.0/stats->unique_kanji);
+    fprintf(output,"hiragana     %9ld\t(%4.1f%%)\n", stats->hiragana_cnt, stats->hiragana_cnt*total_chars);
+    fprintf(output,"katakana     %9ld\t(%4.1f%%)\n", stats->katakana_cnt, stats->katakana_cnt*total_chars);
+    fprintf(output,"Arabic       %9ld\t(%4.1f%%)\n", stats->number_cnt, stats->number_cnt*total_chars);
+    fprintf(output,"Roman        %9ld\t(%4.1f%%)\n", stats->roman_cnt, stats->roman_cnt*total_chars);
+    fprintf(output,"Cyrillic     %9ld\t(%4.1f%%)\n", stats->cyrillic_cnt, stats->cyrillic_cnt*total_chars);
+    fprintf(output,"Greek        %9ld\t(%4.1f%%)\n", stats->greek_cnt, stats->greek_cnt*total_chars);
+    fprintf(output,"graphic      %9ld\t(%4.1f%%)\n", stats->graphic_cnt, stats->graphic_cnt*total_chars);
+    fprintf(output,"punctuation  %9ld\t(%4.1f%%)\n", stats->punctuation_cnt, stats->punctuation_cnt*total_chars);
+    fprintf(output,"symbols      %9ld\t(%4.1f%%)\n", stats->symbol_cnt, stats->symbol_cnt*total_chars);
+    fprintf(output,"space        %9ld\t(%4.1f%%)\n", stats->space_cnt, stats->space_cnt*total_chars);
+    fprintf(output,"other        %9ld\t(%4.1f%%)\n", stats->other_cnt, stats->other_cnt*total_chars);
   }
 
   /*
@@ -465,7 +465,7 @@ show_fractions( output, kanji_vector, stats )
       kanji_count++;
       element++;
     }
-    fprintf(output,"   %d0              %5d\n",(int)fraction, (int)kanji_count);
+    fprintf(output,"   %d0              %5ld\n", fraction, kanji_count);
     target += ten_percent;
   }
 
@@ -477,9 +477,9 @@ show_fractions( output, kanji_vector, stats )
       kanji_count++;
       element++;
     }
-    fprintf(output,"   9%d              %5d\n", (int)fraction, (int)kanji_count);
+    fprintf(output,"   9%d              %5ld\n", fraction, kanji_count);
     target += one_percent;
   }
-  fprintf(output,"  100              %5d\n", (int)stats->unique_kanji);
+  fprintf(output,"  100              %5ld\n", stats->unique_kanji);
 } /* show_fractions() */
